change type from int to size_t, explicit code style

+++ inc/parser_utils.h
@@ -6,7 +6,7 @@
 /*   By: aautin <aautin@student.42.fr >             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/11 15:30:29 by pnguyen-          #+#    #+#             */
-/*   Updated: 2024/03/25 14:50:05 by aautin           ###   ########.fr       */
+/*   Updated: 2024/05/15 17:21:25 by aautin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,11 +20,11 @@ typedef struct s_expansion
 	int				ignore_quotes;
 }	t_expansion;
 
-int		nbr_len(unsigned char nbr);
+size_t	nbr_len(unsigned char nbr);
 void	nbr_data(char **data, unsigned char nbr);
 int		change_quote_mode(char data, int *mode);
-int		pathname_len(char *pathname);
-int		expand_len(t_expansion const *config, char data[]);
+size_t	pathname_len(char *pathname);
+size_t	expand_len(t_expansion const *config, char data[]);
 void	expand_data(t_expansion const *config, char data[], char newdata[]);
 
 #endif
